[YUNIKORN-258] fix make lint to really run (#144)
The make target for lint in the shim only works if there are uncommitted
changes. In all other cases the check will pass. That means the PR
pre-commit check in travis will never fail.
Fix the lint target to perform the same check iwith the same settings as
the core.

Fixes: #144